refactor: extract TimelineItem into its own file
