+++ frontend/src/components/timeline/TimelineItem.tsx
@@ -1,4 +1,4 @@
-import { TimelineEvent } from "@/utils/mockTimelineEvents";
+import { TimelineEvent } from "@/model/TimelineEvent";
 import { formatStringList } from "@/utils/stringUtils";
 
 interface Props {
