+++ frontend/src/model/TimelineEvent.ts
@@ -1,0 +1,8 @@
+export interface TimelineEvent {
+  id: number;
+  name: string;
+  date: Date;
+  desc: string | null;
+  contacts: string[];
+  categories: string[];
+}
