+++ frontend/src/service/TempServerFacadeImpl.ts
@@ -1,11 +1,12 @@
 import IServerFacade from "@/service/IServerFacade";
 import { mockContacts } from "@/utils/mockContacts";
-import { mockTimelineEvents, TimelineEvent } from "@/utils/mockTimelineEvents";
+import { mockTimelineEvents } from "@/utils/mockTimelineEvents";
 import { ContactListQuery, ContactListQueryResult, NewContactData } from "@/service/ContactService";
 import { NewEventData, TimelineQuery, TimelineQueryResult } from "@/service/EventService";
 import { Category } from "@/model/Category";
 import { AuthToken } from "@/model/AuthToken";
 import { Contact } from "@/model/Contact";
+import { TimelineEvent } from "@/model/TimelineEvent";
 
 const CATEGORY: Category = { id: 0, name: "CATEGORY-NAME" };
